#7 separate internal and external calls to better serve proxy pattern
