--- service-proxy.gox
@@ -1,14 +1,0 @@
-package pkg
-
-import "fmt"
-
-// Invoke ...
-func Invoke() string {
-	// TODO:load up the sub grpc Services proxy
-
-	fmt.Println("In sys-share GRPC Services Proxy")
-
-	return "hello from sys-share GRPC Services Proxy"
-
-	
-}
