--- sys-account/service/flu/lib/api/v2/authn.pb.dart
@@ -1,550 +1,0 @@
-///
-//  Generated code. Do not modify.
-//  source: authn.proto
-//
-// @dart = 2.3
-// ignore_for_file: camel_case_types,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type
-
-import 'dart:core' as $core;
-
-import 'package:protobuf/protobuf.dart' as $pb;
-
-import 'google/protobuf/timestamp.pb.dart' as $2;
-
-class ErrorReason extends $pb.GeneratedMessage {
-  static final $pb.BuilderInfo _i = $pb.BuilderInfo('ErrorReason', package: const $pb.PackageName('getcouragenow.sys.v2.sys_account'), createEmptyInstance: create)
-    ..aOS(1, 'reason')
-    ..hasRequiredFields = false
-  ;
-
-  ErrorReason._() : super();
-  factory ErrorReason() => create();
-  factory ErrorReason.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory ErrorReason.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
-  ErrorReason clone() => ErrorReason()..mergeFromMessage(this);
-  ErrorReason copyWith(void Function(ErrorReason) updates) => super.copyWith((message) => updates(message as ErrorReason));
-  $pb.BuilderInfo get info_ => _i;
-  @$core.pragma('dart2js:noInline')
-  static ErrorReason create() => ErrorReason._();
-  ErrorReason createEmptyInstance() => create();
-  static $pb.PbList<ErrorReason> createRepeated() => $pb.PbList<ErrorReason>();
-  @$core.pragma('dart2js:noInline')
-  static ErrorReason getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<ErrorReason>(create);
-  static ErrorReason _defaultInstance;
-
-  @$pb.TagNumber(1)
-  $core.String get reason => $_getSZ(0);
-  @$pb.TagNumber(1)
-  set reason($core.String v) { $_setString(0, v); }
-  @$pb.TagNumber(1)
-  $core.bool hasReason() => $_has(0);
-  @$pb.TagNumber(1)
-  void clearReason() => clearField(1);
-}
-
-class RegisterRequest extends $pb.GeneratedMessage {
-  static final $pb.BuilderInfo _i = $pb.BuilderInfo('RegisterRequest', package: const $pb.PackageName('getcouragenow.sys.v2.sys_account'), createEmptyInstance: create)
-    ..aOS(1, 'email')
-    ..aOS(2, 'password')
-    ..aOS(3, 'passwordConfirm')
-    ..hasRequiredFields = false
-  ;
-
-  RegisterRequest._() : super();
-  factory RegisterRequest() => create();
-  factory RegisterRequest.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory RegisterRequest.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
-  RegisterRequest clone() => RegisterRequest()..mergeFromMessage(this);
-  RegisterRequest copyWith(void Function(RegisterRequest) updates) => super.copyWith((message) => updates(message as RegisterRequest));
-  $pb.BuilderInfo get info_ => _i;
-  @$core.pragma('dart2js:noInline')
-  static RegisterRequest create() => RegisterRequest._();
-  RegisterRequest createEmptyInstance() => create();
-  static $pb.PbList<RegisterRequest> createRepeated() => $pb.PbList<RegisterRequest>();
-  @$core.pragma('dart2js:noInline')
-  static RegisterRequest getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<RegisterRequest>(create);
-  static RegisterRequest _defaultInstance;
-
-  @$pb.TagNumber(1)
-  $core.String get email => $_getSZ(0);
-  @$pb.TagNumber(1)
-  set email($core.String v) { $_setString(0, v); }
-  @$pb.TagNumber(1)
-  $core.bool hasEmail() => $_has(0);
-  @$pb.TagNumber(1)
-  void clearEmail() => clearField(1);
-
-  @$pb.TagNumber(2)
-  $core.String get password => $_getSZ(1);
-  @$pb.TagNumber(2)
-  set password($core.String v) { $_setString(1, v); }
-  @$pb.TagNumber(2)
-  $core.bool hasPassword() => $_has(1);
-  @$pb.TagNumber(2)
-  void clearPassword() => clearField(2);
-
-  @$pb.TagNumber(3)
-  $core.String get passwordConfirm => $_getSZ(2);
-  @$pb.TagNumber(3)
-  set passwordConfirm($core.String v) { $_setString(2, v); }
-  @$pb.TagNumber(3)
-  $core.bool hasPasswordConfirm() => $_has(2);
-  @$pb.TagNumber(3)
-  void clearPasswordConfirm() => clearField(3);
-}
-
-class LoginRequest extends $pb.GeneratedMessage {
-  static final $pb.BuilderInfo _i = $pb.BuilderInfo('LoginRequest', package: const $pb.PackageName('getcouragenow.sys.v2.sys_account'), createEmptyInstance: create)
-    ..aOS(1, 'email')
-    ..aOS(2, 'password')
-    ..hasRequiredFields = false
-  ;
-
-  LoginRequest._() : super();
-  factory LoginRequest() => create();
-  factory LoginRequest.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory LoginRequest.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
-  LoginRequest clone() => LoginRequest()..mergeFromMessage(this);
-  LoginRequest copyWith(void Function(LoginRequest) updates) => super.copyWith((message) => updates(message as LoginRequest));
-  $pb.BuilderInfo get info_ => _i;
-  @$core.pragma('dart2js:noInline')
-  static LoginRequest create() => LoginRequest._();
-  LoginRequest createEmptyInstance() => create();
-  static $pb.PbList<LoginRequest> createRepeated() => $pb.PbList<LoginRequest>();
-  @$core.pragma('dart2js:noInline')
-  static LoginRequest getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<LoginRequest>(create);
-  static LoginRequest _defaultInstance;
-
-  @$pb.TagNumber(1)
-  $core.String get email => $_getSZ(0);
-  @$pb.TagNumber(1)
-  set email($core.String v) { $_setString(0, v); }
-  @$pb.TagNumber(1)
-  $core.bool hasEmail() => $_has(0);
-  @$pb.TagNumber(1)
-  void clearEmail() => clearField(1);
-
-  @$pb.TagNumber(2)
-  $core.String get password => $_getSZ(1);
-  @$pb.TagNumber(2)
-  set password($core.String v) { $_setString(1, v); }
-  @$pb.TagNumber(2)
-  $core.bool hasPassword() => $_has(1);
-  @$pb.TagNumber(2)
-  void clearPassword() => clearField(2);
-}
-
-class LoginResponse extends $pb.GeneratedMessage {
-  static final $pb.BuilderInfo _i = $pb.BuilderInfo('LoginResponse', package: const $pb.PackageName('getcouragenow.sys.v2.sys_account'), createEmptyInstance: create)
-    ..aOB(1, 'success')
-    ..aOS(2, 'accessToken')
-    ..aOS(3, 'refreshToken')
-    ..aOM<ErrorReason>(4, 'errorReason', subBuilder: ErrorReason.create)
-    ..aOM<$2.Timestamp>(5, 'lastLogin', subBuilder: $2.Timestamp.create)
-    ..hasRequiredFields = false
-  ;
-
-  LoginResponse._() : super();
-  factory LoginResponse() => create();
-  factory LoginResponse.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory LoginResponse.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
-  LoginResponse clone() => LoginResponse()..mergeFromMessage(this);
-  LoginResponse copyWith(void Function(LoginResponse) updates) => super.copyWith((message) => updates(message as LoginResponse));
-  $pb.BuilderInfo get info_ => _i;
-  @$core.pragma('dart2js:noInline')
-  static LoginResponse create() => LoginResponse._();
-  LoginResponse createEmptyInstance() => create();
-  static $pb.PbList<LoginResponse> createRepeated() => $pb.PbList<LoginResponse>();
-  @$core.pragma('dart2js:noInline')
-  static LoginResponse getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<LoginResponse>(create);
-  static LoginResponse _defaultInstance;
-
-  @$pb.TagNumber(1)
-  $core.bool get success => $_getBF(0);
-  @$pb.TagNumber(1)
-  set success($core.bool v) { $_setBool(0, v); }
-  @$pb.TagNumber(1)
-  $core.bool hasSuccess() => $_has(0);
-  @$pb.TagNumber(1)
-  void clearSuccess() => clearField(1);
-
-  @$pb.TagNumber(2)
-  $core.String get accessToken => $_getSZ(1);
-  @$pb.TagNumber(2)
-  set accessToken($core.String v) { $_setString(1, v); }
-  @$pb.TagNumber(2)
-  $core.bool hasAccessToken() => $_has(1);
-  @$pb.TagNumber(2)
-  void clearAccessToken() => clearField(2);
-
-  @$pb.TagNumber(3)
-  $core.String get refreshToken => $_getSZ(2);
-  @$pb.TagNumber(3)
-  set refreshToken($core.String v) { $_setString(2, v); }
-  @$pb.TagNumber(3)
-  $core.bool hasRefreshToken() => $_has(2);
-  @$pb.TagNumber(3)
-  void clearRefreshToken() => clearField(3);
-
-  @$pb.TagNumber(4)
-  ErrorReason get errorReason => $_getN(3);
-  @$pb.TagNumber(4)
-  set errorReason(ErrorReason v) { setField(4, v); }
-  @$pb.TagNumber(4)
-  $core.bool hasErrorReason() => $_has(3);
-  @$pb.TagNumber(4)
-  void clearErrorReason() => clearField(4);
-  @$pb.TagNumber(4)
-  ErrorReason ensureErrorReason() => $_ensure(3);
-
-  @$pb.TagNumber(5)
-  $2.Timestamp get lastLogin => $_getN(4);
-  @$pb.TagNumber(5)
-  set lastLogin($2.Timestamp v) { setField(5, v); }
-  @$pb.TagNumber(5)
-  $core.bool hasLastLogin() => $_has(4);
-  @$pb.TagNumber(5)
-  void clearLastLogin() => clearField(5);
-  @$pb.TagNumber(5)
-  $2.Timestamp ensureLastLogin() => $_ensure(4);
-}
-
-class RegisterResponse extends $pb.GeneratedMessage {
-  static final $pb.BuilderInfo _i = $pb.BuilderInfo('RegisterResponse', package: const $pb.PackageName('getcouragenow.sys.v2.sys_account'), createEmptyInstance: create)
-    ..aOB(1, 'success')
-    ..aOS(2, 'successMsg')
-    ..aOM<ErrorReason>(3, 'errorReason', subBuilder: ErrorReason.create)
-    ..hasRequiredFields = false
-  ;
-
-  RegisterResponse._() : super();
-  factory RegisterResponse() => create();
-  factory RegisterResponse.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory RegisterResponse.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
-  RegisterResponse clone() => RegisterResponse()..mergeFromMessage(this);
-  RegisterResponse copyWith(void Function(RegisterResponse) updates) => super.copyWith((message) => updates(message as RegisterResponse));
-  $pb.BuilderInfo get info_ => _i;
-  @$core.pragma('dart2js:noInline')
-  static RegisterResponse create() => RegisterResponse._();
-  RegisterResponse createEmptyInstance() => create();
-  static $pb.PbList<RegisterResponse> createRepeated() => $pb.PbList<RegisterResponse>();
-  @$core.pragma('dart2js:noInline')
-  static RegisterResponse getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<RegisterResponse>(create);
-  static RegisterResponse _defaultInstance;
-
-  @$pb.TagNumber(1)
-  $core.bool get success => $_getBF(0);
-  @$pb.TagNumber(1)
-  set success($core.bool v) { $_setBool(0, v); }
-  @$pb.TagNumber(1)
-  $core.bool hasSuccess() => $_has(0);
-  @$pb.TagNumber(1)
-  void clearSuccess() => clearField(1);
-
-  @$pb.TagNumber(2)
-  $core.String get successMsg => $_getSZ(1);
-  @$pb.TagNumber(2)
-  set successMsg($core.String v) { $_setString(1, v); }
-  @$pb.TagNumber(2)
-  $core.bool hasSuccessMsg() => $_has(1);
-  @$pb.TagNumber(2)
-  void clearSuccessMsg() => clearField(2);
-
-  @$pb.TagNumber(3)
-  ErrorReason get errorReason => $_getN(2);
-  @$pb.TagNumber(3)
-  set errorReason(ErrorReason v) { setField(3, v); }
-  @$pb.TagNumber(3)
-  $core.bool hasErrorReason() => $_has(2);
-  @$pb.TagNumber(3)
-  void clearErrorReason() => clearField(3);
-  @$pb.TagNumber(3)
-  ErrorReason ensureErrorReason() => $_ensure(2);
-}
-
-class ForgotPasswordRequest extends $pb.GeneratedMessage {
-  static final $pb.BuilderInfo _i = $pb.BuilderInfo('ForgotPasswordRequest', package: const $pb.PackageName('getcouragenow.sys.v2.sys_account'), createEmptyInstance: create)
-    ..aOS(1, 'email')
-    ..hasRequiredFields = false
-  ;
-
-  ForgotPasswordRequest._() : super();
-  factory ForgotPasswordRequest() => create();
-  factory ForgotPasswordRequest.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory ForgotPasswordRequest.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
-  ForgotPasswordRequest clone() => ForgotPasswordRequest()..mergeFromMessage(this);
-  ForgotPasswordRequest copyWith(void Function(ForgotPasswordRequest) updates) => super.copyWith((message) => updates(message as ForgotPasswordRequest));
-  $pb.BuilderInfo get info_ => _i;
-  @$core.pragma('dart2js:noInline')
-  static ForgotPasswordRequest create() => ForgotPasswordRequest._();
-  ForgotPasswordRequest createEmptyInstance() => create();
-  static $pb.PbList<ForgotPasswordRequest> createRepeated() => $pb.PbList<ForgotPasswordRequest>();
-  @$core.pragma('dart2js:noInline')
-  static ForgotPasswordRequest getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<ForgotPasswordRequest>(create);
-  static ForgotPasswordRequest _defaultInstance;
-
-  @$pb.TagNumber(1)
-  $core.String get email => $_getSZ(0);
-  @$pb.TagNumber(1)
-  set email($core.String v) { $_setString(0, v); }
-  @$pb.TagNumber(1)
-  $core.bool hasEmail() => $_has(0);
-  @$pb.TagNumber(1)
-  void clearEmail() => clearField(1);
-}
-
-class ForgotPasswordResponse extends $pb.GeneratedMessage {
-  static final $pb.BuilderInfo _i = $pb.BuilderInfo('ForgotPasswordResponse', package: const $pb.PackageName('getcouragenow.sys.v2.sys_account'), createEmptyInstance: create)
-    ..aOB(1, 'success')
-    ..aOS(2, 'successMsg')
-    ..aOM<ErrorReason>(3, 'errorReason', subBuilder: ErrorReason.create)
-    ..aOM<$2.Timestamp>(4, 'forgotPasswordRequestedAt', subBuilder: $2.Timestamp.create)
-    ..hasRequiredFields = false
-  ;
-
-  ForgotPasswordResponse._() : super();
-  factory ForgotPasswordResponse() => create();
-  factory ForgotPasswordResponse.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory ForgotPasswordResponse.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
-  ForgotPasswordResponse clone() => ForgotPasswordResponse()..mergeFromMessage(this);
-  ForgotPasswordResponse copyWith(void Function(ForgotPasswordResponse) updates) => super.copyWith((message) => updates(message as ForgotPasswordResponse));
-  $pb.BuilderInfo get info_ => _i;
-  @$core.pragma('dart2js:noInline')
-  static ForgotPasswordResponse create() => ForgotPasswordResponse._();
-  ForgotPasswordResponse createEmptyInstance() => create();
-  static $pb.PbList<ForgotPasswordResponse> createRepeated() => $pb.PbList<ForgotPasswordResponse>();
-  @$core.pragma('dart2js:noInline')
-  static ForgotPasswordResponse getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<ForgotPasswordResponse>(create);
-  static ForgotPasswordResponse _defaultInstance;
-
-  @$pb.TagNumber(1)
-  $core.bool get success => $_getBF(0);
-  @$pb.TagNumber(1)
-  set success($core.bool v) { $_setBool(0, v); }
-  @$pb.TagNumber(1)
-  $core.bool hasSuccess() => $_has(0);
-  @$pb.TagNumber(1)
-  void clearSuccess() => clearField(1);
-
-  @$pb.TagNumber(2)
-  $core.String get successMsg => $_getSZ(1);
-  @$pb.TagNumber(2)
-  set successMsg($core.String v) { $_setString(1, v); }
-  @$pb.TagNumber(2)
-  $core.bool hasSuccessMsg() => $_has(1);
-  @$pb.TagNumber(2)
-  void clearSuccessMsg() => clearField(2);
-
-  @$pb.TagNumber(3)
-  ErrorReason get errorReason => $_getN(2);
-  @$pb.TagNumber(3)
-  set errorReason(ErrorReason v) { setField(3, v); }
-  @$pb.TagNumber(3)
-  $core.bool hasErrorReason() => $_has(2);
-  @$pb.TagNumber(3)
-  void clearErrorReason() => clearField(3);
-  @$pb.TagNumber(3)
-  ErrorReason ensureErrorReason() => $_ensure(2);
-
-  @$pb.TagNumber(4)
-  $2.Timestamp get forgotPasswordRequestedAt => $_getN(3);
-  @$pb.TagNumber(4)
-  set forgotPasswordRequestedAt($2.Timestamp v) { setField(4, v); }
-  @$pb.TagNumber(4)
-  $core.bool hasForgotPasswordRequestedAt() => $_has(3);
-  @$pb.TagNumber(4)
-  void clearForgotPasswordRequestedAt() => clearField(4);
-  @$pb.TagNumber(4)
-  $2.Timestamp ensureForgotPasswordRequestedAt() => $_ensure(3);
-}
-
-class ResetPasswordRequest extends $pb.GeneratedMessage {
-  static final $pb.BuilderInfo _i = $pb.BuilderInfo('ResetPasswordRequest', package: const $pb.PackageName('getcouragenow.sys.v2.sys_account'), createEmptyInstance: create)
-    ..aOS(1, 'email')
-    ..aOS(2, 'password')
-    ..aOS(3, 'passwordConfirm')
-    ..hasRequiredFields = false
-  ;
-
-  ResetPasswordRequest._() : super();
-  factory ResetPasswordRequest() => create();
-  factory ResetPasswordRequest.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory ResetPasswordRequest.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
-  ResetPasswordRequest clone() => ResetPasswordRequest()..mergeFromMessage(this);
-  ResetPasswordRequest copyWith(void Function(ResetPasswordRequest) updates) => super.copyWith((message) => updates(message as ResetPasswordRequest));
-  $pb.BuilderInfo get info_ => _i;
-  @$core.pragma('dart2js:noInline')
-  static ResetPasswordRequest create() => ResetPasswordRequest._();
-  ResetPasswordRequest createEmptyInstance() => create();
-  static $pb.PbList<ResetPasswordRequest> createRepeated() => $pb.PbList<ResetPasswordRequest>();
-  @$core.pragma('dart2js:noInline')
-  static ResetPasswordRequest getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<ResetPasswordRequest>(create);
-  static ResetPasswordRequest _defaultInstance;
-
-  @$pb.TagNumber(1)
-  $core.String get email => $_getSZ(0);
-  @$pb.TagNumber(1)
-  set email($core.String v) { $_setString(0, v); }
-  @$pb.TagNumber(1)
-  $core.bool hasEmail() => $_has(0);
-  @$pb.TagNumber(1)
-  void clearEmail() => clearField(1);
-
-  @$pb.TagNumber(2)
-  $core.String get password => $_getSZ(1);
-  @$pb.TagNumber(2)
-  set password($core.String v) { $_setString(1, v); }
-  @$pb.TagNumber(2)
-  $core.bool hasPassword() => $_has(1);
-  @$pb.TagNumber(2)
-  void clearPassword() => clearField(2);
-
-  @$pb.TagNumber(3)
-  $core.String get passwordConfirm => $_getSZ(2);
-  @$pb.TagNumber(3)
-  set passwordConfirm($core.String v) { $_setString(2, v); }
-  @$pb.TagNumber(3)
-  $core.bool hasPasswordConfirm() => $_has(2);
-  @$pb.TagNumber(3)
-  void clearPasswordConfirm() => clearField(3);
-}
-
-class ResetPasswordResponse extends $pb.GeneratedMessage {
-  static final $pb.BuilderInfo _i = $pb.BuilderInfo('ResetPasswordResponse', package: const $pb.PackageName('getcouragenow.sys.v2.sys_account'), createEmptyInstance: create)
-    ..aOB(1, 'success')
-    ..aOS(2, 'successMsg')
-    ..aOM<ErrorReason>(3, 'errorReason', subBuilder: ErrorReason.create)
-    ..aOM<$2.Timestamp>(4, 'resetPasswordRequestedAt', subBuilder: $2.Timestamp.create)
-    ..hasRequiredFields = false
-  ;
-
-  ResetPasswordResponse._() : super();
-  factory ResetPasswordResponse() => create();
-  factory ResetPasswordResponse.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory ResetPasswordResponse.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
-  ResetPasswordResponse clone() => ResetPasswordResponse()..mergeFromMessage(this);
-  ResetPasswordResponse copyWith(void Function(ResetPasswordResponse) updates) => super.copyWith((message) => updates(message as ResetPasswordResponse));
-  $pb.BuilderInfo get info_ => _i;
-  @$core.pragma('dart2js:noInline')
-  static ResetPasswordResponse create() => ResetPasswordResponse._();
-  ResetPasswordResponse createEmptyInstance() => create();
-  static $pb.PbList<ResetPasswordResponse> createRepeated() => $pb.PbList<ResetPasswordResponse>();
-  @$core.pragma('dart2js:noInline')
-  static ResetPasswordResponse getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<ResetPasswordResponse>(create);
-  static ResetPasswordResponse _defaultInstance;
-
-  @$pb.TagNumber(1)
-  $core.bool get success => $_getBF(0);
-  @$pb.TagNumber(1)
-  set success($core.bool v) { $_setBool(0, v); }
-  @$pb.TagNumber(1)
-  $core.bool hasSuccess() => $_has(0);
-  @$pb.TagNumber(1)
-  void clearSuccess() => clearField(1);
-
-  @$pb.TagNumber(2)
-  $core.String get successMsg => $_getSZ(1);
-  @$pb.TagNumber(2)
-  set successMsg($core.String v) { $_setString(1, v); }
-  @$pb.TagNumber(2)
-  $core.bool hasSuccessMsg() => $_has(1);
-  @$pb.TagNumber(2)
-  void clearSuccessMsg() => clearField(2);
-
-  @$pb.TagNumber(3)
-  ErrorReason get errorReason => $_getN(2);
-  @$pb.TagNumber(3)
-  set errorReason(ErrorReason v) { setField(3, v); }
-  @$pb.TagNumber(3)
-  $core.bool hasErrorReason() => $_has(2);
-  @$pb.TagNumber(3)
-  void clearErrorReason() => clearField(3);
-  @$pb.TagNumber(3)
-  ErrorReason ensureErrorReason() => $_ensure(2);
-
-  @$pb.TagNumber(4)
-  $2.Timestamp get resetPasswordRequestedAt => $_getN(3);
-  @$pb.TagNumber(4)
-  set resetPasswordRequestedAt($2.Timestamp v) { setField(4, v); }
-  @$pb.TagNumber(4)
-  $core.bool hasResetPasswordRequestedAt() => $_has(3);
-  @$pb.TagNumber(4)
-  void clearResetPasswordRequestedAt() => clearField(4);
-  @$pb.TagNumber(4)
-  $2.Timestamp ensureResetPasswordRequestedAt() => $_ensure(3);
-}
-
-class RefreshAccessTokenRequest extends $pb.GeneratedMessage {
-  static final $pb.BuilderInfo _i = $pb.BuilderInfo('RefreshAccessTokenRequest', package: const $pb.PackageName('getcouragenow.sys.v2.sys_account'), createEmptyInstance: create)
-    ..aOS(1, 'refreshToken')
-    ..hasRequiredFields = false
-  ;
-
-  RefreshAccessTokenRequest._() : super();
-  factory RefreshAccessTokenRequest() => create();
-  factory RefreshAccessTokenRequest.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory RefreshAccessTokenRequest.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
-  RefreshAccessTokenRequest clone() => RefreshAccessTokenRequest()..mergeFromMessage(this);
-  RefreshAccessTokenRequest copyWith(void Function(RefreshAccessTokenRequest) updates) => super.copyWith((message) => updates(message as RefreshAccessTokenRequest));
-  $pb.BuilderInfo get info_ => _i;
-  @$core.pragma('dart2js:noInline')
-  static RefreshAccessTokenRequest create() => RefreshAccessTokenRequest._();
-  RefreshAccessTokenRequest createEmptyInstance() => create();
-  static $pb.PbList<RefreshAccessTokenRequest> createRepeated() => $pb.PbList<RefreshAccessTokenRequest>();
-  @$core.pragma('dart2js:noInline')
-  static RefreshAccessTokenRequest getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<RefreshAccessTokenRequest>(create);
-  static RefreshAccessTokenRequest _defaultInstance;
-
-  @$pb.TagNumber(1)
-  $core.String get refreshToken => $_getSZ(0);
-  @$pb.TagNumber(1)
-  set refreshToken($core.String v) { $_setString(0, v); }
-  @$pb.TagNumber(1)
-  $core.bool hasRefreshToken() => $_has(0);
-  @$pb.TagNumber(1)
-  void clearRefreshToken() => clearField(1);
-}
-
-class RefreshAccessTokenResponse extends $pb.GeneratedMessage {
-  static final $pb.BuilderInfo _i = $pb.BuilderInfo('RefreshAccessTokenResponse', package: const $pb.PackageName('getcouragenow.sys.v2.sys_account'), createEmptyInstance: create)
-    ..aOS(1, 'accessToken')
-    ..aOM<ErrorReason>(2, 'errorReason', subBuilder: ErrorReason.create)
-    ..hasRequiredFields = false
-  ;
-
-  RefreshAccessTokenResponse._() : super();
-  factory RefreshAccessTokenResponse() => create();
-  factory RefreshAccessTokenResponse.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
-  factory RefreshAccessTokenResponse.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
-  RefreshAccessTokenResponse clone() => RefreshAccessTokenResponse()..mergeFromMessage(this);
-  RefreshAccessTokenResponse copyWith(void Function(RefreshAccessTokenResponse) updates) => super.copyWith((message) => updates(message as RefreshAccessTokenResponse));
-  $pb.BuilderInfo get info_ => _i;
-  @$core.pragma('dart2js:noInline')
-  static RefreshAccessTokenResponse create() => RefreshAccessTokenResponse._();
-  RefreshAccessTokenResponse createEmptyInstance() => create();
-  static $pb.PbList<RefreshAccessTokenResponse> createRepeated() => $pb.PbList<RefreshAccessTokenResponse>();
-  @$core.pragma('dart2js:noInline')
-  static RefreshAccessTokenResponse getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<RefreshAccessTokenResponse>(create);
-  static RefreshAccessTokenResponse _defaultInstance;
-
-  @$pb.TagNumber(1)
-  $core.String get accessToken => $_getSZ(0);
-  @$pb.TagNumber(1)
-  set accessToken($core.String v) { $_setString(0, v); }
-  @$pb.TagNumber(1)
-  $core.bool hasAccessToken() => $_has(0);
-  @$pb.TagNumber(1)
-  void clearAccessToken() => clearField(1);
-
-  @$pb.TagNumber(2)
-  ErrorReason get errorReason => $_getN(1);
-  @$pb.TagNumber(2)
-  set errorReason(ErrorReason v) { setField(2, v); }
-  @$pb.TagNumber(2)
-  $core.bool hasErrorReason() => $_has(1);
-  @$pb.TagNumber(2)
-  void clearErrorReason() => clearField(2);
-  @$pb.TagNumber(2)
-  ErrorReason ensureErrorReason() => $_ensure(1);
-}
-
