--- sys-account/service/flu/lib/api/v2/authn.pbjson.dart
@@ -1,102 +1,0 @@
-///
-//  Generated code. Do not modify.
-//  source: authn.proto
-//
-// @dart = 2.3
-// ignore_for_file: camel_case_types,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type
-
-const ErrorReason$json = const {
-  '1': 'ErrorReason',
-  '2': const [
-    const {'1': 'reason', '3': 1, '4': 1, '5': 9, '10': 'reason'},
-  ],
-};
-
-const RegisterRequest$json = const {
-  '1': 'RegisterRequest',
-  '2': const [
-    const {'1': 'email', '3': 1, '4': 1, '5': 9, '10': 'email'},
-    const {'1': 'password', '3': 2, '4': 1, '5': 9, '10': 'password'},
-    const {'1': 'password_confirm', '3': 3, '4': 1, '5': 9, '10': 'passwordConfirm'},
-  ],
-};
-
-const LoginRequest$json = const {
-  '1': 'LoginRequest',
-  '2': const [
-    const {'1': 'email', '3': 1, '4': 1, '5': 9, '10': 'email'},
-    const {'1': 'password', '3': 2, '4': 1, '5': 9, '10': 'password'},
-  ],
-};
-
-const LoginResponse$json = const {
-  '1': 'LoginResponse',
-  '2': const [
-    const {'1': 'success', '3': 1, '4': 1, '5': 8, '10': 'success'},
-    const {'1': 'access_token', '3': 2, '4': 1, '5': 9, '10': 'accessToken'},
-    const {'1': 'refresh_token', '3': 3, '4': 1, '5': 9, '10': 'refreshToken'},
-    const {'1': 'error_reason', '3': 4, '4': 1, '5': 11, '6': '.getcouragenow.sys.v2.sys_account.ErrorReason', '10': 'errorReason'},
-    const {'1': 'last_login', '3': 5, '4': 1, '5': 11, '6': '.google.protobuf.Timestamp', '10': 'lastLogin'},
-  ],
-};
-
-const RegisterResponse$json = const {
-  '1': 'RegisterResponse',
-  '2': const [
-    const {'1': 'success', '3': 1, '4': 1, '5': 8, '10': 'success'},
-    const {'1': 'success_msg', '3': 2, '4': 1, '5': 9, '10': 'successMsg'},
-    const {'1': 'error_reason', '3': 3, '4': 1, '5': 11, '6': '.getcouragenow.sys.v2.sys_account.ErrorReason', '10': 'errorReason'},
-  ],
-};
-
-const ForgotPasswordRequest$json = const {
-  '1': 'ForgotPasswordRequest',
-  '2': const [
-    const {'1': 'email', '3': 1, '4': 1, '5': 9, '10': 'email'},
-  ],
-};
-
-const ForgotPasswordResponse$json = const {
-  '1': 'ForgotPasswordResponse',
-  '2': const [
-    const {'1': 'success', '3': 1, '4': 1, '5': 8, '10': 'success'},
-    const {'1': 'success_msg', '3': 2, '4': 1, '5': 9, '10': 'successMsg'},
-    const {'1': 'error_reason', '3': 3, '4': 1, '5': 11, '6': '.getcouragenow.sys.v2.sys_account.ErrorReason', '10': 'errorReason'},
-    const {'1': 'forgot_password_requested_at', '3': 4, '4': 1, '5': 11, '6': '.google.protobuf.Timestamp', '10': 'forgotPasswordRequestedAt'},
-  ],
-};
-
-const ResetPasswordRequest$json = const {
-  '1': 'ResetPasswordRequest',
-  '2': const [
-    const {'1': 'email', '3': 1, '4': 1, '5': 9, '10': 'email'},
-    const {'1': 'password', '3': 2, '4': 1, '5': 9, '10': 'password'},
-    const {'1': 'password_confirm', '3': 3, '4': 1, '5': 9, '10': 'passwordConfirm'},
-  ],
-};
-
-const ResetPasswordResponse$json = const {
-  '1': 'ResetPasswordResponse',
-  '2': const [
-    const {'1': 'success', '3': 1, '4': 1, '5': 8, '10': 'success'},
-    const {'1': 'success_msg', '3': 2, '4': 1, '5': 9, '10': 'successMsg'},
-    const {'1': 'error_reason', '3': 3, '4': 1, '5': 11, '6': '.getcouragenow.sys.v2.sys_account.ErrorReason', '10': 'errorReason'},
-    const {'1': 'reset_password_requested_at', '3': 4, '4': 1, '5': 11, '6': '.google.protobuf.Timestamp', '10': 'resetPasswordRequestedAt'},
-  ],
-};
-
-const RefreshAccessTokenRequest$json = const {
-  '1': 'RefreshAccessTokenRequest',
-  '2': const [
-    const {'1': 'refresh_token', '3': 1, '4': 1, '5': 9, '10': 'refreshToken'},
-  ],
-};
-
-const RefreshAccessTokenResponse$json = const {
-  '1': 'RefreshAccessTokenResponse',
-  '2': const [
-    const {'1': 'access_token', '3': 1, '4': 1, '5': 9, '10': 'accessToken'},
-    const {'1': 'error_reason', '3': 2, '4': 1, '5': 11, '6': '.getcouragenow.sys.v2.sys_account.ErrorReason', '10': 'errorReason'},
-  ],
-};
-
