+++ sys-account/service/flu/lib/api/v2/services.pb.dart
@@ -1,7 +1,9 @@
 ///
 //  Generated code. Do not modify.
-//  source: authn.proto
+//  source: services.proto
 //
 // @dart = 2.3
 // ignore_for_file: camel_case_types,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type
+
+import 'dart:core' as $core;
