+++ sys-account/service/flu/lib/api/v2/services.pbjson.dart
@@ -1,0 +1,7 @@
+///
+//  Generated code. Do not modify.
+//  source: services.proto
+//
+// @dart = 2.3
+// ignore_for_file: camel_case_types,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type
+
